refactor(dkg): move dkg util method definitions onto the DKG structs
