+++ sn_node/src/node/dkg/mod.rs
@@ -6,9 +6,8 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-pub(crate) mod dkg_msgs_utils;
 mod session;
 mod voter;
 
-pub(crate) use self::{dkg_msgs_utils::DkgFailureSigSetUtils, voter::DkgVoter};
+pub(crate) use self::voter::DkgVoter;
 pub(crate) use sn_interface::messaging::system::{KeyedSig, SigShare};
