+++ sn_node/src/node/dkg/session.rs
@@ -8,7 +8,6 @@
 
 use crate::node::{
     api::cmds::{next_timer_token, Cmd},
-    dkg::dkg_msgs_utils::{DkgFailureSigSetUtils, DkgFailureSigUtils},
     messages::WireMsgUtils,
     Result,
 };
